Add basic styles to filter toolbar
